Add navigation guard via user store
