--- src/common/store.js
@@ -1,16 +1,0 @@
-import {reactive, readonly} from 'vue';
-
-export class Store {
-    constructor() {
-      let data = this.data()
-      this.setup(data)
-      this.state = reactive(data)
-    }
-  
-    setup(data) {}
-  
-    getState() {
-      return readonly(this.state)
-    }
-  }
-
